requirements/testing.txt
factory-boy==2.12.0
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-isort==0.3.1
pytest-mock==1.10.4
pytest-sugar==0.9.2
pytest==4.4.1
